requirements.txt
requests>=2.31.0
websockets>=12.0
cryptography>=41.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
numpy>=1.24.0
scipy>=1.10.0
